devices: proxy: do not acknowledge write commands
Write accessess cannot fail (in the CommandResult sense) and the result
did not carry any data, so remove the response from the Write command.

This should improve the speed of write requests for sandboxed devices.

For example, with the sandboxed serial device, boot time with a release
build of crosvm on my workstation goes from 1.7 seconds to 1.2 seconds,
measured by timing a boot with a missing init so that the kernel panics
and shuts down immediately.

BUG=None
TEST=time crosvm run -p init=bogus vm_kernel

Change-Id: I125bb831235ca741ae1cc6c86a02a5d863d1a211
Signed-off-by: Daniel Verkamp <dverkamp@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/crosvm/+/1853970
Reviewed-by: Zach Reizner <zachr@chromium.org>
Tested-by: kokoro <noreply+kokoro@google.com>